Crate bdk_persist

BDK Persist

This crate is home to the PersistBackend trait which defines the behavior of a database to perform the task of persisting changes made to BDK data structures.

The Persist type provides a convenient wrapper around a PersistBackend that allows staging changes before committing them.

Structs

CombinedChangeSet

Changes from a combination of bdk_chain structures.

Persist

Persist wraps a PersistBackend to create a convenient staging area for changes (C) before they are persisted.

Traits

PersistBackend

A persistence backend for Persist.