Struct bdk_persist::Persist

pub struct Persist<C> { /* private fields */ }

Persist wraps a PersistBackend to create a convenient staging area for changes (C) before they are persisted.

Not all changes to the in-memory representation needs to be written to disk right away, so Persist::stage can be used to stage changes first and then Persist::commit can be used to write changes to disk.

Implementations

impl<C> Persist<C>
where C: Default + Append,

pub fn new(backend: impl PersistBackend<C> + Send + Sync + 'static) -> Self

Create a new Persist from PersistBackend.

pub fn stage(&mut self, changeset: C)

Stage a changeset to be committed later with commit.

pub fn staged(&self) -> &C

Get the changes that have not been committed yet.

pub fn commit(&mut self) -> Result<Option<C>>

Commit the staged changes to the underlying persistence backend.

Changes that are committed (if any) are returned.

Error

Returns a backend-defined error if this fails.

pub fn stage_and_commit(&mut self, changeset: C) -> Result<Option<C>>

Stages a new changeset and commits it (along with any other previously staged changes) to the persistence backend

Convenience method for calling stage and then commit.